+++ packages/toujou-accordion/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
-All notable changes to the `@toujou/location-accordion` package will be documented in this file
+All notable changes to the `@toujou/toujou-accordion` package will be documented in this file
+
+## [2.1.0] - 2025-07-03
+- Implement keyboard interactions to comply with WAI ARIA guidelines
 
 ## [2.0.29] - 2023-09-26
 - refactor the component to stop using decorators